re-structured contract
